+++ packages/expo-background-fetch/src/ExpoBackgroundFetch.ts
@@ -1,0 +1,3 @@
+import { NativeModulesProxy } from '@unimodules/core';
+
+export default NativeModulesProxy.ExpoBackgroundFetch;
