+++ packages/expo-background-fetch/src/ExpoBackgroundFetch.web.ts
@@ -1,0 +1,10 @@
+import { BackgroundFetchStatus } from './BackgroundFetch.types';
+
+export default {
+  get name(): string {
+    return 'ExpoBackgroundFetch';
+  },
+  async getStatusAsync(): Promise<BackgroundFetchStatus | null> {
+    return BackgroundFetchStatus.Restricted;
+  },
+};
